Add files via upload
This file creates a list of wordType objects (an object with a word and the point value of the word according to Scrabble's scoring system)
from a the official TWLO6 Scrabble tournament official wordlist (all 178,691 words), which are then used in scrabbleCalculator.py to calculate all the possible playable words
given a certain set of tiles. In addition to printing every playable word to console, it also list the point value in front of it.