Add files via upload
This file is a slightly modified version of scrabbleCalculator.py, which is designed to address the original program's assumption that you have an infinite amount of each tile. This program is supposed to create a modifiable replica of your hand, which can then be used to judge whether or not you have a sufficient amount of each letter to spell a word from the word list.